requirements.txt
fastapi==0.95.0
asyncpg==0.27.0
psycopg2-binary==2.9.5
SQLAlchemy==2.0.15
alembic==1.11.2
redis==4.5.4
pyjwt==2.6.0
passlib==1.7.4
uvicorn==0.22.0
python-jose
redis>=4.0.0
